Revert "[Light Switch] Switch desktop wallpapers with the Light/Dark mode (#4…"
This reverts commit 19c9b4e1fd23579b22a3778e39e60ce9bfa81f16.
