+++ .github/actions/spell-check/expect.txt
@@ -457,7 +457,6 @@ DWMWINDOWATTRIBUTE
 DWMWINDOWMAXIMIZEDCHANGE
 DWORDLONG
 dworigin
-DWPOS
 dwrite
 dxgi
 eab
@@ -678,7 +677,6 @@ hicon
 HICONSM
 HIDEREADONLY
 HIDEWINDOW
-hif
 Hif
 HIMAGELIST
 himl
@@ -840,7 +838,6 @@ jpe
 jpnime
 Jsons
 jsonval
-jxl
 jxr
 keybd
 KEYBDDATA
@@ -1467,7 +1464,6 @@ recyclebin
 Redist
 Reencode
 REFCLSID
-REFGUID
 REFIID
 REGCLS
 regfile
@@ -1574,7 +1570,6 @@ SETBUDDYINT
 SETCONTEXT
 SETCURSEL
 setcursor
-SETDESKWALLPAPER
 SETFOCUS
 SETFOREGROUND
 SETHOTKEY
@@ -1663,7 +1658,6 @@ SKEXP
 SKIPOWNPROCESS
 sku
 SLGP
-slideshow
 sln
 slnf
 slnx
@@ -1902,7 +1896,6 @@ unwide
 unzoom
 UOffset
 UOI
-UPDATEINIFILE
 UPDATENOW
 UPDATEREGISTRY
 updown
