--- src/modules/LightSwitch/LightSwitchCommon/SettingsConstants.h
@@ -1,23 +1,0 @@
-#pragma once
-
-enum class SettingId
-{
-    ScheduleMode = 0,
-    Latitude,
-    Longitude,
-    LightTime,
-    DarkTime,
-    Sunrise_Offset,
-    Sunset_Offset,
-    ChangeSystem,
-    ChangeApps,
-    WallpaperEnabled,
-    WallpaperVirtualDesktopEnabled,
-    WallpaperStyleLight,
-    WallpaperStyleDark,
-    WallpaperPathLight,
-    WallpaperPathDark
-};
-
-inline constexpr wchar_t PERSONALIZATION_REGISTRY_PATH[] = L"Software\\Microsoft\\Windows\\CurrentVersion\\Themes\\Personalize";
-inline constexpr wchar_t NIGHT_LIGHT_REGISTRY_PATH[] = L"Software\\Microsoft\\Windows\\CurrentVersion\\CloudStore\\Store\\DefaultAccount\\Current\\default$windows.data.bluelightreduction.bluelightreductionstate\\windows.data.bluelightreduction.bluelightreductionstate";
