+++ src/modules/LightSwitch/LightSwitchModuleInterface/ThemeHelper.cpp
@@ -1,0 +1,106 @@
+#include "pch.h"
+#include <windows.h>
+#include "ThemeHelper.h"
+
+// Controls changing the themes.
+static void ResetColorPrevalence()
+{
+    HKEY hKey;
+    if (RegOpenKeyEx(HKEY_CURRENT_USER,
+                     L"Software\\Microsoft\\Windows\\CurrentVersion\\Themes\\Personalize",
+                     0,
+                     KEY_SET_VALUE,
+                     &hKey) == ERROR_SUCCESS)
+    {
+        DWORD value = 0; // back to default value
+        RegSetValueEx(hKey, L"ColorPrevalence", 0, REG_DWORD, reinterpret_cast<const BYTE*>(&value), sizeof(value));
+        RegCloseKey(hKey);
+
+        SendMessageTimeout(HWND_BROADCAST, WM_SETTINGCHANGE, 0, reinterpret_cast<LPARAM>(L"ImmersiveColorSet"), SMTO_ABORTIFHUNG, 5000, nullptr);
+
+        SendMessageTimeout(HWND_BROADCAST, WM_THEMECHANGED, 0, 0, SMTO_ABORTIFHUNG, 5000, nullptr);
+
+        SendMessageTimeout(HWND_BROADCAST, WM_DWMCOLORIZATIONCOLORCHANGED, 0, 0, SMTO_ABORTIFHUNG, 5000, nullptr);
+    }
+}
+
+void SetAppsTheme(bool mode)
+{
+    HKEY hKey;
+    if (RegOpenKeyEx(HKEY_CURRENT_USER,
+                     L"Software\\Microsoft\\Windows\\CurrentVersion\\Themes\\Personalize",
+                     0,
+                     KEY_SET_VALUE,
+                     &hKey) == ERROR_SUCCESS)
+    {
+        DWORD value = mode;
+        RegSetValueEx(hKey, L"AppsUseLightTheme", 0, REG_DWORD, reinterpret_cast<const BYTE*>(&value), sizeof(value));
+        RegCloseKey(hKey);
+
+        SendMessageTimeout(HWND_BROADCAST, WM_SETTINGCHANGE, 0, reinterpret_cast<LPARAM>(L"ImmersiveColorSet"), SMTO_ABORTIFHUNG, 5000, nullptr);
+
+        SendMessageTimeout(HWND_BROADCAST, WM_THEMECHANGED, 0, 0, SMTO_ABORTIFHUNG, 5000, nullptr);
+    }
+}
+
+void SetSystemTheme(bool mode)
+{
+    HKEY hKey;
+    if (RegOpenKeyEx(HKEY_CURRENT_USER,
+                     L"Software\\Microsoft\\Windows\\CurrentVersion\\Themes\\Personalize",
+                     0,
+                     KEY_SET_VALUE,
+                     &hKey) == ERROR_SUCCESS)
+    {
+        DWORD value = mode;
+        RegSetValueEx(hKey, L"SystemUsesLightTheme", 0, REG_DWORD, reinterpret_cast<const BYTE*>(&value), sizeof(value));
+        RegCloseKey(hKey);
+
+        if (mode) // if are changing to light mode
+        {
+            ResetColorPrevalence();
+        }
+
+        SendMessageTimeout(HWND_BROADCAST, WM_SETTINGCHANGE, 0, reinterpret_cast<LPARAM>(L"ImmersiveColorSet"), SMTO_ABORTIFHUNG, 5000, nullptr);
+
+        SendMessageTimeout(HWND_BROADCAST, WM_THEMECHANGED, 0, 0, SMTO_ABORTIFHUNG, 5000, nullptr);
+    }
+}
+
+bool GetCurrentSystemTheme()
+{
+    HKEY hKey;
+    DWORD value = 1; // default = light
+    DWORD size = sizeof(value);
+
+    if (RegOpenKeyEx(HKEY_CURRENT_USER,
+                     L"Software\\Microsoft\\Windows\\CurrentVersion\\Themes\\Personalize",
+                     0,
+                     KEY_READ,
+                     &hKey) == ERROR_SUCCESS)
+    {
+        RegQueryValueEx(hKey, L"SystemUsesLightTheme", nullptr, nullptr, reinterpret_cast<LPBYTE>(&value), &size);
+        RegCloseKey(hKey);
+    }
+
+    return value == 1; // true = light, false = dark
+}
+
+bool GetCurrentAppsTheme()
+{
+    HKEY hKey;
+    DWORD value = 1;
+    DWORD size = sizeof(value);
+
+    if (RegOpenKeyEx(HKEY_CURRENT_USER,
+                     L"Software\\Microsoft\\Windows\\CurrentVersion\\Themes\\Personalize",
+                     0,
+                     KEY_READ,
+                     &hKey) == ERROR_SUCCESS)
+    {
+        RegQueryValueEx(hKey, L"AppsUseLightTheme", nullptr, nullptr, reinterpret_cast<LPBYTE>(&value), &size);
+        RegCloseKey(hKey);
+    }
+
+    return value == 1; // true = light, false = dark
+}
