+++ src/modules/LightSwitch/LightSwitchModuleInterface/ThemeHelper.h
@@ -1,0 +1,5 @@
+#pragma once
+void SetSystemTheme(bool dark);
+void SetAppsTheme(bool dark);
+bool GetCurrentSystemTheme();
+bool GetCurrentAppsTheme();
