+++ src/modules/LightSwitch/LightSwitchService/LightSwitchService.vcxproj
@@ -53,7 +53,18 @@
       <ConformanceMode>true</ConformanceMode>
       <PrecompiledHeader>NotUsing</PrecompiledHeader>
       <PreprocessorDefinitions>%(PreprocessorDefinitions)</PreprocessorDefinitions>
-      <AdditionalIncludeDirectories>..\LightSwitchCommon;.\..\;..\..\..\common;..\..\..\common\logger;..\..\..\common\utils;..\..\..\common\SettingsAPI;..\..\..\common\Telemetry;..\..\..\;..\..\..\..\deps\spdlog\include;./;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>
+        ./../;
+        ..\..\..\common;
+        ..\..\..\common\logger;
+        ..\..\..\common\utils;
+        ..\..\..\common\SettingsAPI;
+        ..\..\..\common\Telemetry;
+        ..\..\..\;
+        ..\..\..\..\deps\spdlog\include;
+        ./;
+        %(AdditionalIncludeDirectories)
+      </AdditionalIncludeDirectories>
     </ClCompile>
     <Link>
       <SubSystem>Windows</SubSystem>
@@ -66,7 +77,8 @@
     <ClCompile Include="LightSwitchSettings.cpp" />
     <ClCompile Include="LightSwitchStateManager.cpp" />
     <ClCompile Include="NightLightRegistryObserver.cpp" />
-    <ClCompile Include="..\LightSwitchCommon\ThemeHelper.cpp" />
+    <ClCompile Include="SettingsConstants.cpp" />
+    <ClCompile Include="ThemeHelper.cpp" />
     <ClCompile Include="ThemeScheduler.cpp" />
     <ClCompile Include="trace.cpp" />
     <ClCompile Include="WinHookEventIDs.cpp" />
@@ -79,9 +91,9 @@
     <ClInclude Include="LightSwitchStateManager.h" />
     <ClInclude Include="LightSwitchUtils.h" />
     <ClInclude Include="NightLightRegistryObserver.h" />
+    <ClInclude Include="SettingsConstants.h" />
     <ClInclude Include="SettingsObserver.h" />
-    <ClInclude Include="..\LightSwitchCommon\ThemeHelper.h" />
-    <ClInclude Include="..\LightSwitchCommon\SettingsConstants.h" />
+    <ClInclude Include="ThemeHelper.h" />
     <ClInclude Include="ThemeScheduler.h" />
     <ClInclude Include="trace.h" />
     <ClInclude Include="WinHookEventIDs.h" />
