+++ src/modules/LightSwitch/LightSwitchService/LightSwitchService.vcxproj.filters
@@ -21,10 +21,13 @@
     <ClCompile Include="ThemeScheduler.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="..\LightSwitchCommon\ThemeHelper.cpp">
+    <ClCompile Include="ThemeHelper.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="LightSwitchSettings.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="SettingsConstants.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="WinHookEventIDs.cpp">
@@ -47,10 +50,10 @@
     <ClInclude Include="ThemeHelper.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\LightSwitchCommon\LightSwitchSettings.h">
+    <ClInclude Include="LightSwitchSettings.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\LightSwitchCommon\SettingsConstants.h">
+    <ClInclude Include="SettingsConstants.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="SettingsObserver.h">
