+++ src/modules/LightSwitch/LightSwitchService/LightSwitchSettings.cpp
@@ -253,66 +253,6 @@ void LightSwitchSettings::LoadSettings()
         {
             Trace::LightSwitch::ThemeTargetChanged(m_settings.changeApps, m_settings.changeSystem);
         }
-
-        if (const auto jsonVal = values.get_bool_value(L"wallpaperEnabled"))
-        {
-            auto val = *jsonVal;
-            if (m_settings.wallpaperEnabled != val)
-            {
-                m_settings.wallpaperEnabled = val;
-                NotifyObservers(SettingId::WallpaperEnabled);
-            }
-        }
-
-        if (const auto jsonVal = values.get_bool_value(L"wallpaperVirtualDesktopEnabled"))
-        {
-            auto val = *jsonVal;
-            if (m_settings.wallpaperVirtualDesktop != val)
-            {
-                m_settings.wallpaperVirtualDesktop = val;
-                NotifyObservers(SettingId::WallpaperVirtualDesktopEnabled);
-            }
-        }
-
-        if (const auto jsonVal = values.get_int_value(L"wallpaperStyleLight"))
-        {
-            auto val = *jsonVal;
-            if (m_settings.wallpaperStyleLight != val)
-            {
-                m_settings.wallpaperStyleLight = val;
-                NotifyObservers(SettingId::WallpaperStyleLight);
-            }
-        }
-
-        if (const auto jsonVal = values.get_int_value(L"wallpaperStyleDark"))
-        {
-            auto val = *jsonVal;
-            if (m_settings.wallpaperStyleDark != val)
-            {
-                m_settings.wallpaperStyleDark = val;
-                NotifyObservers(SettingId::WallpaperStyleDark);
-            }
-        }
-
-        if (const auto jsonVal = values.get_string_value(L"wallpaperPathLight"))
-        {
-            auto val = *jsonVal;
-            if (m_settings.wallpaperPathLight != val)
-            {
-                m_settings.wallpaperPathLight = val;
-                NotifyObservers(SettingId::WallpaperPathLight);
-            }
-        }
-
-        if (const auto jsonVal = values.get_string_value(L"wallpaperPathDark"))
-        {
-            auto val = *jsonVal;
-            if (m_settings.wallpaperPathDark != val)
-            {
-                m_settings.wallpaperPathDark = val;
-                NotifyObservers(SettingId::WallpaperPathDark);
-            }
-        }
     }
     catch (...)
     {
