+++ src/modules/LightSwitch/LightSwitchService/SettingsConstants.cpp
@@ -1,0 +1,1 @@
+#include "SettingsConstants.h"
