+++ src/modules/LightSwitch/LightSwitchService/ThemeHelper.h
@@ -1,0 +1,6 @@
+#pragma once
+void SetSystemTheme(bool dark);
+void SetAppsTheme(bool dark);
+bool GetCurrentSystemTheme();
+bool GetCurrentAppsTheme();
+bool IsNightLightEnabled();
